backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pandas
torch
transformers
Pillow
tqdm
python-multipart
faiss-cpu
matplotlib
google-genai